gcppractice/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
google-cloud-firestore
mysql-connector-python